missions/W2/M5/team/requirements.txt
selenium
webdriver-manager
pandas
PyYAML
beautifulsoup4
readability-lxml
requests
python-dotenv
playwright
wordcloud
kiwipiepy
matplotlib
ko-ww-stopwords
